Update README.md
SmartHire Functional Overview:

SmartHire is a decentralized, Peer-to-Peer recruitment platform where candidates
are evaluated and ranked by their technical peers before being matched with fitting
job opportunities by a data-driven matching algorithm.

Each candidate will have a public and private profile, with all contact information
(Full name, E-mail, Phone number, Address, etc.) being initially withheld from any
third parties. The public profile will display a summary of their skills, experience and
education as well as any other relevant information that the candidate chooses to
share with the public.

Candidates will then be matched with a technical peer (someone of similar or greater
level of experience in the industry and skill category of the candidate) that will
conduct a Techical Peer Review (TPR) to take place over a phone or video call,
where they will ask the candidate questions to evaluate and rate the candidate’s
skills and select from a variety of recommended open positions that they believe the
candidate would be a good fit for.

In return, the candidate will also evaluate and give feedback to the Technical Peer
on their performance during the peer review, which will serve as a reputation builder
for the technical peer (more on that to follow).

Both the candidate and the technical peer will receive SmartCoins for investing their
time and providing value to the ecosystem, with more incentives to come should the
interaction lead to a positive outcome in the future (more on that to follow).

If the said candidate is interested in exploring the opportunities that he or she was
matched with, they can show their interest to the companies by applying to those
jobs. If they are not currently looking, they can hit the “snooze” button on the
matched opportunities, and the matching algorithm will automatically notify them of
similar positions that become available at their selected interval of time (1 month, 3
months, 6 months, 12 months).

If at any point in time the candidate wants to “refresh” their skills evaluation (perhaps
they have gained valuable experience with a new technology, or worked on a side
project that exposed them to a new industry), they can always request an additional
Techical Peer Review (TPR) only this time they would have to spend some of their 
SmartCoin to compensate the Technical Peer for their time (Each candidate will
receive one FREE review when they join the platform).

Once the candidate is ready to explore new opportunities, they will be able to further
explore the jobs and companies of their choice. Several possible use cases emerge
to help them do that:

a. They can simply indicate that they are interested in the role to the employer
by flipping the “interested” switch on the job, and wait to see if the employer
will request further information based on their profile.

b. If the employer wants to see their resume, they will have the option to offer
the candidate SmartCoin for sharing it (candidates will be able to set their
preferences including the cost of sharing their information, doing phone and
in-person interviews, and technical assessments, etc.)

c. If the employer wants to schedule a phone interview based on the resume,
they will have the option to offer the candidate SmartCoin for agreeing to
the interview.

d. If the employer wants to schedule a follow-up in person interview, they will
have the option to offer the candidate SmartCoin for coming to the
interview.

e. If the employer wants the candidate to do a technical assessment, they will
have the option to offer the candidate SmartCoin for taking it.

f. If the employer wants to offer the candidate a job, they will have the option
to offer a SmartCoin signing bonus to the employee for agreeing to join their
company. The system off course isn’t perfect, and not every candidate will be matched with the
opportunities and companies that they would like to explore. SmartCoin provides
several interesting use cases here also:

g. The candidate can browse the jobs and companies of their choice and
request an information interview with the companies they like, offering to
pay in SmartCoin for gaining a deeper insight into the company and seeing 
if it’s the right fit for them (Companies can control the cost or availability of
information interviews in their settings)

h. The candidate can select to complete any of the available “challenges” that
a company offers to interested candidates in order to better assess their
skills. (Companies can setup these challenges in their settings as part of
their talent acquisition strategy)

i. The candidate can apply to complete some freelance work for their
company of choice (Companies can setup some redundant or especially
difficult problems as part of their talent acquisition strategy)

Once each step of the interview process is completed, companies will be required to
provide feedback to candidates, for which they will be compensated in SmartCoin.
When the company has made their decision to hire the candidate, they will generate a
Smart Contract which will transfer a selected amount of SmartCoin to the candidate
upon their acceptance of the offer of employment (which can be completed offline).

Along with the candidate, the company will also pay SmartCoins to all the technical
peers who have done a Techical Peer Review, and a special referral SmartCoin bonus
to the Peer who recommended the candidate for this particular opportunity.